# Replace manual use of `(~B_c*(1/(B_c*~B_c)[0])` with a call to normalInv