Since `normalInv` does extra checks, this also introduces a flag to disable those extra checks.

This also speeds up `inv()` - it no longer performs the multiplication `self * ~self` twice.