@eric-wieser 2019-11-07
Timings before:
```python
In [1]: from clifford.g3c import *

In [2]: a = (e1 + e3 + e123)

In [3]: %timeit a.normalInv()
133 µs ± 5.08 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [4]: %timeit a.inv()
252 µs ± 19.7 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [5]: %timeit ~a*(1/(a*~a)[0])
38.9 µs ± 1.82 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

Timings after:
```python
In [1]: from clifford.g3c import *

In [2]: a = (e1 + e3 + e123)

In [3]: %timeit a.normalInv(check=False)
34 µs ± 1.05 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [4]: %timeit a.normalInv()
131 µs ± 5.79 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [5]: %timeit a.inv()
139 µs ± 10.4 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

`.normalInv(check=False)` is actually faster than the full expression, since it doesn't compute `~a` twice!